Move Kill and PrepareForChildren to platform specific implementation
